06/01/2020 - Fightmode v4
Meilleur agencement pour ne rien faire en cas de click hors d'une zone autorisée.

+++ Game_files/Godot/HOMM_FightMode_4/Characters/Temporary.gd
@@ -12,7 +12,7 @@ func _ready():
 		
 # warning-ignore:unused_argument
 func _input(event):
-	if Input.is_action_just_pressed("ui_leftclic") :
+	if Input.is_action_just_pressed("ui_leftclic") && displacement_allowed ==true :
 		Delete_all_displacement_tiles()
 	displacement_allowed = false
